Day_01/02_PRACTICAL TASK/02_openrouter/story_generater_ai/pyproject.toml
[project]
name = "ai-weather-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
